Revert "refactor(entities): move entity platforms to entities/ package"
This reverts commit 55d67859538eaa616accca684f50d4de4b1b323d.
